Transaction confirmation moved from modal to page.
