+++ app/components/commitfunds/SignInstructionGeneric.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
 export const SignInstructionGeneric: React.SFC = () =>
-  <div>Please confirm transaction using signer</div>;
+  <div>Please confirm transaction using signer and wait for it to be send to Ethereum node.</div>;
 
 export default SignInstructionGeneric;
